Add support for encrypting audio
